Create ts_utils.py
Supporting utilities to run the segmentation metrics. This code load the TotalSegmenator code once, and keeps it in memory. This allows the evaluation of a batch of images using one single model. Note that this code was written with the intent to run in a Docker container and is archived here.